Change internal cuts to mechanical holes; turn MOSFETs back-to-back
Replaced the internal cuts with NPTH mechanical holes for the
panelized snap line.

Swapped the MOSFET locations and rotated one so that they are now
back-to-back with the heat sink toward the middle to leave more
options for the K2 LOAD connector.  This should allow wires,
headers, or 0.1" pitch screw connector terminal block connectors.