Replace mutually exclusive MOSFETs with single MOSFET
Previously, there were two MOSFETs in the schematic, one for each pin
configuration, to make it easy to place two components on the board.
These have been replaced by a single MOSFET and an edited modules with
two sets of pins, one for each of the GDS and DGS pinouts.